// src/app/Models/Clientes/motivoRetiro.model.ts
export class MotivoRetiroModel {
    IdTercero: number = 0;
    FechaRetiro: string = "";
    IdAsesor: number = 0;
    DescripcionOtro: string = "";
    MotivoRetiro: number = 0;
}
export class EnvioMotivoModel {
    Nombre: string = "";
    Documento: string = "";
    CiudadExpedicion: string = "";
    Direccion: string = "";
    Celular: string = "";
    Empresa: string = "";
    Telefono: string = "";
    Email: string = "";
    TelefonoEmpresa: string = "";
    NombreAsesor: string = "";
    Motivo: string = "";
    IdMotivo: string = "";
    Otro: string = "";
}
export class MotivoRetiroJuridicoModel {
    IdRetiroLog: number = 0;
    IdJuridico: number = 0;
    FechaRetiro: string = "";
    IdAsesor: number = 0;
    Descripcion: string = "";
    MotivoRetiro: number = 0;
}
export class EnvioMotivoJuridicoModel {
    Nombre: string = "";
    Documento: string = "";
    CiudadExpedicion: string = "";
    Direccion: string = "";
    Celular: string = "";
    Empresa: string = "";
    Telefono: string = "";
    Email: string = "";
    TelefonoEmpresa: string = "";
    NombreAsesor: string = "";
    Motivo: string = "";
    IdMotivo: string = "";
    Otro: string = "";
}